Added repeating notification for start of event. (2)

+++ phoneProfilesPlus/src/main/java/sk/henrichg/phoneprofilesplus/StartEventNotificationBroadcastReceiver.java
@@ -59,7 +59,7 @@ public class StartEventNotificationBroadcastReceiver extends BroadcastReceiver {
             Intent intent = new Intent(context, StartEventNotificationBroadcastReceiver.class);
             intent.putExtra(PPApplication.EXTRA_EVENT_ID, event._id);
 
-            PendingIntent pendingIntent = PendingIntent.getBroadcast(context.getApplicationContext(), (int) 0, intent, PendingIntent.FLAG_CANCEL_CURRENT);
+            PendingIntent pendingIntent = PendingIntent.getBroadcast(context.getApplicationContext(), 0, intent, PendingIntent.FLAG_CANCEL_CURRENT);
 
             AlarmManager alarmManager = (AlarmManager) context.getSystemService(Activity.ALARM_SERVICE);
             if (alarmManager != null) {
